- Setting the classpath for Reflections SubTypesScanner. - Corrected test package name to lower case - Enhanced the package unit test to test deeper hierarchies
