+++ jsonschema-maven-plugin/src/test/resources/reference-test-cases/TestClassC-reference.json
@@ -1,0 +1,12 @@
+{
+  "$schema" : "http://json-schema.org/draft-07/schema#",
+  "type" : "object",
+  "properties" : {
+    "aDouble" : {
+      "type" : "number"
+    },
+    "anotherDouble" : {
+      "type" : "number"
+    }
+  }
+}
