+++ jsonschema-maven-plugin/src/test/resources/reference-test-cases/TwoClasses-pom.xml
@@ -6,8 +6,8 @@
                 <artifactId>jsonschema-maven-plugin</artifactId>
                 <configuration>
                     <classNames>
-                        <className>com.github.victools.jsonschema.plugin.maven.TestPackage.TestClassA</className>
-                        <className>com.github.victools.jsonschema.plugin.maven.TestPackage.TestClassB</className>
+                        <className>com.github.victools.jsonschema.plugin.maven.testpackage.TestClassA</className>
+                        <className>com.github.victools.jsonschema.plugin.maven.testpackage.TestClassB</className>
                     </classNames>
                     <schemaFilePath>target/generated-test-sources/TwoClasses</schemaFilePath>
                     <schemaFileName>{0}.schema</schemaFileName>
